Fixed bug:
In the event that the widget window size changes such that it is no longer enclosed entirely within the bounds of the monitor, the application will now adjust the widget's position such that it again becomes enclosed entirely within the bounds of the monitor.

An increase in widget window size may be instigated either when the widget displays additional text, or when the user increases its font size.